+++ model/Article.php
@@ -4,10 +4,8 @@ namespace P5\Model;
 // Call BDD
 require_once ("../model/Manager.php");
 
-class PostManager extends Manager
+class Article extends Manager
 {
-	
-
 	public function listArticles()
     {
         $db = $this->dbConnect();
@@ -27,16 +25,6 @@ class PostManager extends Manager
 
         return $article;
     } 
-
-    public function listComments($idArticle)
-    {
-    	$db = $this->dbConnect();
-    	$req = $db->prepare("SELECT post_id, author, content, DATE_FORMAT(comment_date, '%d-%m-%Y à %Hh%i') AS comment_date, authorized  FROM comment WHERE post_id=? ORDER BY comment_date ASC");
-    	$req->execute(array($idArticle));
-    	$comments = $req->fetchAll();
-
-    	return $comments;
-    }
 
     public function updateArticle($updateTitle, $updateContent, $updateAuthor, $idArticle)
     {
@@ -67,17 +55,6 @@ class PostManager extends Manager
 
         return $author;
     }
-
-    public function postComment($idArticle, $authorComment, $content)
-    {
-        $db = $this->dbConnect();
-        $req = $db->prepare('INSERT INTO comment (post_id, author, content, comment_date, authorized) VALUES(?, ?, ?, NOW(), 0)');
-        $newComment = $req->execute(array($idArticle, $authorComment, $content));
-
-        return $newComment;
-    }
-
-
 
 }
